Hhbt65 (#313)
* ADDED: User new Account Flag, Online Status Flag

* FIXED: User status coming offline always

* CHANGED: default durations for online and recent

* ADDED: Clear Chat Room Feature

* CHANGED: delete room and get rooms based on user

* FIXED:error in api

* FIXED: Last message on cleared chat

* FIXED: Latest message on Cleared Chat

* ADDED: v2 for Get Chat Rooms

* ADDED: Filter Limit with different City Selected

* Filter Limit to 1 with city selected

* ADDED:Colleges CRUD

* FIXED: Duplicate Route name in API

* ADDED: Add College API and Get List API

* ADDED: Save College in New Flow

* FIXED: Filters Applied Bug

* ADDED: College search and my profile

* ADDED: College search and my profile

* ADDED: Test users flow

* ADDED: Test Users Flow

* CHANGED: Filter Logic

* ADDED: City and State List in Colleges

* ADDED: College Details in Profile Resource

* FIXED: 412 status error in states and cities list

* FIXED: Create Chat Room Api

* FIXED: Sticking buttons in List

* CHANGED: Filter Limit

* CHANGED: College in Profile Percentage

* FIXED: Param error in Users

* ADDED: Hide Profile Feature

* FIXED: Home Feed error

* CHANGED: College Flow for allowing Abbreviations

* FIXED: is_blocked and blocking messages

* FIXED: Hide Error

* CHANGED State Selection in Colleges and renamed District to City

+++ app/Http/Controllers/api/v1/BlockController.php
@@ -41,6 +41,7 @@ class BlockController extends Controller
                     $block_profile = BlockUser::firstOrCreate([
                         'block_by'      =>  $auth_id,
                         'blocked_to'    =>  $block_user->id ?? NULL,
+                        'block_type'    =>  'block'
                     ], [
                         'custom_id'     =>  getUniqueString('block_users'),
                     ]);
@@ -67,13 +68,14 @@ class BlockController extends Controller
                             ]
                         ]);
                     } else {
-                        $this->response['meta']['message']  =   rans('api.block.fail');
+                        $this->response['meta']['message']  =  trans('api.block.fail');
                         $this->response['meta']['is_ban'] = false;
                         $this->status = Response::HTTP_NOT_FOUND;
                     }
                 }
-                /* Unblock Profile */ elseif ($request->status == 'unblock') {
-                    $block_profile = BlockUser::whereBlockBy($auth_id)->whereBlockedTo($block_user->id)->firstOrFail();
+                /* Unblock Profile */
+                elseif ($request->status == 'unblock') {
+                    $block_profile = BlockUser::blockedOnly()->whereBlockBy($auth_id)->whereBlockedTo($block_user->id)->firstOrFail();
                     $unblock = $block_profile->delete();
 
                     // Unblock Chat
@@ -98,7 +100,58 @@ class BlockController extends Controller
                             ]
                         ]);
                     } else {
-                        $this->response['meta']['message']  =   rans('api.unblock.fail');
+                        $this->response['meta']['message']  =   trans('api.unblock.fail');
+                        $this->response['meta']['is_ban'] = false;
+                        $this->status = Response::HTTP_NOT_FOUND;
+                    }
+                }
+                elseif ($request->status == 'hide') {
+                    $hide_profile = BlockUser::firstOrCreate([
+                        'block_by'      =>  $auth_id,
+                        'blocked_to'    =>  $block_user->id ?? NULL,
+                        'block_type'    =>  'hide'
+                    ], [
+                        'custom_id'     =>  getUniqueString('block_users'),
+                    ]);
+
+                    DB::commit();
+                    if ($hide_profile->save()) {
+                        $this->status = Response::HTTP_OK;
+                        return ([
+                            'data'  =>  NULL,
+                            'meta' => [
+                                'url'       =>  url()->current(),
+                                'api'       =>  $this->getVersion(),
+                                'language'  =>  app()->getLocale(),
+                                'is_ban'    =>  false,
+                                'message'   =>  trans('api.hide.success'),
+                            ]
+                        ]);
+                    } else {
+                        $this->response['meta']['message']  =   trans('api.hide.fail');
+                        $this->response['meta']['is_ban'] = false;
+                        $this->status = Response::HTTP_NOT_FOUND;
+                    }
+                }
+                elseif ($request->status == 'unhide') {
+                    $hide_profile = BlockUser::hiddenOnly()->whereBlockBy($auth_id)->whereBlockedTo($block_user->id)->firstOrFail();
+                    $unhide = $hide_profile->delete();
+
+                    DB::commit();
+                    if ($unhide) {
+                        $this->status = Response::HTTP_OK;
+                        return ([
+                            'data'  =>  NULL,
+                            'meta' => [
+                                'url'       =>  url()->current(),
+                                'api'       =>  $this->getVersion(),
+                                'language'  =>  app()->getLocale(),
+                                'is_ban'    =>  false,
+                                'message'   =>  trans('api.unhide.success'),
+                            ]
+                        ]);
+                    } else {
+                        $this->response['meta']['message']  =  trans('api.unhide.fail');
                         $this->response['meta']['is_ban'] = false;
                         $this->status = Response::HTTP_NOT_FOUND;
                     }
